requirements.txt
django
pillow
djangorestframework
drf-yasg 
psycopg2
django-rest-auth